+++ src/main/gen/com/javampire/openscad/psi/impl/OpenSCADTestExpRefImpl.java
@@ -29,12 +29,12 @@ public class OpenSCADTestExpRefImpl extends OpenSCADResolvableElementImpl implem
 
   @Override
   public PsiElement setName(String newName) {
-    return OpenSCADPsiImplUtil.setName(this, newName);
+    return OpenSCADPsiUtils.setName(this, newName);
   }
 
   @Override
   public PsiElement getNameIdentifier() {
-    return OpenSCADPsiImplUtil.getNameIdentifier(this);
+    return OpenSCADPsiUtils.getNameIdentifier(this);
   }
 
 }