+++ src/main/java/com/javampire/openscad/psi/OpenSCADNamedElement.java
@@ -2,5 +2,8 @@ package com.javampire.openscad.psi;
 
 import com.intellij.psi.PsiNameIdentifierOwner;
 
+/**
+ * Any element that have a name, variable, functions, module, ...
+ */
 public interface OpenSCADNamedElement extends PsiNameIdentifierOwner {
 }