+++ src/main/java/com/javampire/openscad/psi/stub/function/OpenSCADFunctionIndex.java
@@ -6,14 +6,11 @@ import com.intellij.psi.stubs.StringStubIndexExtension;
 import com.intellij.psi.stubs.StubIndex;
 import com.intellij.psi.stubs.StubIndexKey;
 import com.javampire.openscad.psi.OpenSCADFunctionDeclaration;
-import com.javampire.openscad.references.OpenSCADReferenceResolver;
 import org.jetbrains.annotations.NotNull;
 
 import java.util.Collection;
 
-public class OpenSCADFunctionIndex
-        extends StringStubIndexExtension<OpenSCADFunctionDeclaration>
-        implements OpenSCADReferenceResolver {
+public class OpenSCADFunctionIndex extends StringStubIndexExtension<OpenSCADFunctionDeclaration> {
 
     public static final StubIndexKey<String, OpenSCADFunctionDeclaration> FUNCTION = StubIndexKey.createIndexKey("OpenSCAD.function.name");
 
@@ -21,6 +18,10 @@ public class OpenSCADFunctionIndex
 
     public static OpenSCADFunctionIndex getInstance() {
         return INSTANCE;
+    }
+
+    public boolean traceKeyHashToVirtualFileMapping() {
+        return true;
     }
 
     @NotNull
@@ -35,5 +36,4 @@ public class OpenSCADFunctionIndex
                                                        @NotNull final GlobalSearchScope scope) {
         return StubIndex.getElements(getKey(), functionName, project, null, OpenSCADFunctionDeclaration.class);
     }
-
 }