+++ CHANGELOG.md
@@ -12,10 +12,11 @@
 ### Removed
 
 ### Fixed
+- [Issue-47](https://github.com/ldenisey/idea-openscad/issues/47) Builtin functions and modules documentation does not appear
 - [Issue-46](https://github.com/ldenisey/idea-openscad/issues/46) Some keywords are parsed as identifiers
 - [Issue-41](https://github.com/ldenisey/idea-openscad/issues/41) Show preview activation popup only on OpenSCAD projects
 - Bump dependencies
-- Github set-output action migration
+- GitHub set-output action migration
 
 ## 2.3.2
 ### Added